Move charts to a separate directory
