+++ app/javascript/packs/validators/charts/block_history_chart.js
