+++ app/javascript/packs/validators/charts/skipped_slots_chart.js
