+++ app/javascript/packs/validators/charts/vote_history_chart.js
